lib: rename str-timestamp to timeutils
