--- lib/str-timestamp/CMakeLists.txt
@@ -1,9 +1,0 @@
-set(STR_TIMESTAMP_HEADERS
-    str-timestamp/decode.h
-    PARENT_SCOPE)
-
-set(STR_TIMESTAMP_SOURCES
-    str-timestamp/decode.c
-    PARENT_SCOPE)
-
-add_test_subdirectory(tests)
