--- lib/str-timestamp/Makefile.am
@@ -1,11 +1,0 @@
-str_timestampincludedir = ${pkgincludedir}/str-timestamp
-
-EXTRA_DIST += lib/str-timestamp/CMakeLists.txt
-
-str_timestampinclude_HEADERS =	\
-  lib/str-timestamp/decode.h
-
-str_timestamp_sources =		\
- lib/str-timestamp/decode.c
-
-include lib/str-timestamp/tests/Makefile.am
