+++ lib/timeutils/Makefile.am
@@ -1,0 +1,11 @@
+timeutilsincludedir = ${pkgincludedir}/timeutils
+
+EXTRA_DIST += lib/timeutils/CMakeLists.txt
+
+timeutilsinclude_HEADERS =	\
+  lib/timeutils/decode.h
+
+timeutils_sources =		\
+ lib/timeutils/decode.c
+
+include lib/timeutils/tests/Makefile.am
